requirements.txt
pdf2image
alyn
pytesseract
Pillow
opencv-python
numpy
tqdm
requests
pandas
openpyxl
hello